+++ lib/fields/ncm_field.dart
@@ -39,6 +39,7 @@ class NcmField extends ValidatorField {
     super.decoration,
     super.padding,
     super.hintText,
+    super.contentPadding,
     super.prefix,
     super.prefixIcon,
     super.suffix,