+++ lib/fields/password_field.dart
@@ -37,6 +37,7 @@ class PasswordField extends StringField {
     super.decoration,
     super.padding,
     super.hintText,
+    super.contentPadding,
     super.prefix,
     super.prefixIcon,
     super.suffix,