+++ lib/fields/validator_field.dart
@@ -43,6 +43,7 @@ class ValidatorField extends StringField {
     super.decoration,
     super.padding,
     super.hintText,
+    super.contentPadding,
     super.prefix,
     super.prefixIcon,
     super.suffix,